+++ ex04_staged_object/android/app/src/main/kotlin/com/example/ex04_staged_object/MainActivity.kt
@@ -1,0 +1,12 @@
+package com.example.ex04_staged_object
+
+import androidx.annotation.NonNull;
+import io.flutter.embedding.android.FlutterActivity
+import io.flutter.embedding.engine.FlutterEngine
+import io.flutter.plugins.GeneratedPluginRegistrant
+
+class MainActivity: FlutterActivity() {
+    override fun configureFlutterEngine(@NonNull flutterEngine: FlutterEngine) {
+        GeneratedPluginRegistrant.registerWith(flutterEngine);
+    }
+}