+++ ex04_staged_object/lib/src/blocs/page_two_bloc.dart
@@ -1,0 +1,30 @@
+import 'package:frideos/frideos.dart';
+
+class StagedPageTwoBloc {
+  StagedPageTwoBloc() {
+    print('-------StagedPageTwo BLOC--------');
+    receiverWidget.outStream.listen((map) {
+      staged.setStagesMap(map);
+    });
+  }
+
+  final receiverWidget = StreamedMap<int, Stage>(initialData: {});
+  final receiverStr = StreamedValue<String>();
+
+  // Create an instance with the setStagesMap constructor
+  // or use the default constractor and the setStagesMap method
+  final staged = StagedObject();
+
+  final scrollingText = StreamedValue<String>();
+
+  final timerObject = TimerObject();
+
+  void dispose() {
+    print('-------StagedPageTwo BLOC DISPOSE--------');
+    timerObject.dispose();
+    staged.dispose();
+    receiverWidget.dispose();
+    receiverStr.dispose();
+    scrollingText.dispose();
+  }
+}