+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Version 2.0 (14-01-2020)
+- Migration to frideos package 1.0.0
+- Migration to RxDart 0.23.1
+
 ## Version 1.0 (18-05-2019)
 
 - Repository refactoring